Separate error presentation and sanitizer regression suites
